+++ LotusRMS.Models/Viewmodels/Checkout/CheckoutVM.cs
@@ -15,11 +15,11 @@ namespace LotusRMS.Models.Viewmodels.Checkout
         public OrderVm  Order{ get; set; }
         public string DateTime { get; set; }
         public Guid? Customer_Id { get; set; }
-        public float Total { get; set; }
+        public double Total { get; set; }
         public string Customer_Name { get; set; }
         public string Customer_Address { get; set; }
         public string Customer_Contact { get; set; }
-        public float Paid_Amount { get; set; }
+        public double Paid_Amount { get; set; }
         public string Discount_Type { get; set; }
         public float Discount { get; set; }
         public string Payment_Mode { get; set; }
