+++ LotusRMS.Models/Viewmodels/Customer/CustomerVm.cs
@@ -13,7 +13,7 @@ namespace LotusRMS.Models.Viewmodels.Customer
         public string Address { get; set; }
         public string Contact { get; set; }
         public string PanOrVat { get; set; }
-        public float DueAmount { get; set; }
+        public double DueAmount { get; set; }
         public bool Status { get; set; }
     }
 }
